Initial changes suggested by Eamonn
