+++ galasa-ui/messages/de.json
@@ -266,7 +266,6 @@
     "Time": "Zeit",
     "Method": "Methode",
     "searchPlaceholder": "Suchen",
-    "ariaLabel": "ariaLabel",
     "pagination": {
       "backwardText": "Vorherige Seite",
       "forwardText": "Nächste Seite",
