+++ galasa-ui/src/tests/components/test-runs/test-run-details/ArtifactsTab.test.tsx
@@ -12,7 +12,6 @@ import { checkForZosTerminalFolderStructure } from '@/utils/3270/checkFor3270Fol
 import { downloadArtifactFromServer } from '@/actions/runsAction';
 import { handleDownload } from '@/utils/artifacts';
 import { FeatureFlagProvider } from '@/contexts/FeatureFlagContext';
-import { mock } from 'node:test';
 
 // Mock dependencies
 jest.mock('@/actions/runsAction');
@@ -652,95 +651,6 @@ describe('ArtifactsTab', () => {
       await waitFor(() => {
         expect(screen.getByText('Error downloading artifact for run Test Run')).toBeInTheDocument();
       });
-    });
-  });
-
-  describe('Check For Zos Terminal Folder Structure Method', () => {
-    const setZos3270TerminalFolderExists = jest.fn();
-
-    const mockRootFolder: TreeNodeData = {
-      name: '',
-      isFile: false,
-      children: {
-        zos3270: {
-          name: 'zos3270',
-          isFile: false,
-          children: {
-            terminals: {
-              name: 'terminals',
-              isFile: false,
-              children: {
-                GAL91419_1: {
-                  name: 'GAL91419_1',
-                  isFile: false,
-                  children: {
-                    'GAL91419_1-00207.gz': {
-                      name: 'GAL91419_1-00207.gz',
-                      runId: '',
-                      url: '/artifacts/zos3270/terminals/GAL91419_1/GAL91419_1-00207.gz',
-                      isFile: true,
-                      children: {},
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-    };
-
-    const mockEmptyRootFolder: TreeNodeData = {
-      name: 'root',
-      isFile: false,
-      children: {},
-    };
-
-    beforeEach(() => {
-      jest.resetModules();
-      setZos3270TerminalFolderExists.mockReset();
-    });
-
-    test('sets Zos3270TerminalFolderExists to true when the structure is found', () => {
-      checkForZosTerminalFolderStructure(
-        mockRootFolder,
-        setZos3270TerminalFolderExists,
-        mockSetZos3270TerminalData
-      );
-      expect(setZos3270TerminalFolderExists).toHaveBeenCalledWith(true);
-    });
-
-    test('sets Zos3270TerminalFolderExists to false when empty file structure is found', () => {
-      checkForZosTerminalFolderStructure(
-        mockEmptyRootFolder,
-        setZos3270TerminalFolderExists,
-        mockSetZos3270TerminalData
-      );
-      expect(setZos3270TerminalFolderExists).toHaveBeenCalledWith(false);
-    });
-
-    test('sets Zos3270TerminalFolderExists to false when "zos3270" folder does not contain a "terminals" subfolder', () => {
-      (mockRootFolder.children as { [key: string]: any })['zos3270'] = {
-        children: {},
-      };
-      checkForZosTerminalFolderStructure(
-        mockRootFolder,
-        setZos3270TerminalFolderExists,
-        mockSetZos3270TerminalData
-      );
-      expect(setZos3270TerminalFolderExists).toHaveBeenCalledWith(false);
-    });
-
-    test('sets Zos3270TerminalFolderExists to false when "zos3270/terminals" is not populated with one or more files', () => {
-      (mockRootFolder.children.zos3270.children as { [key: string]: any })['terminals'] = {
-        children: {},
-      };
-      checkForZosTerminalFolderStructure(
-        mockRootFolder,
-        setZos3270TerminalFolderExists,
-        mockSetZos3270TerminalData
-      );
-      expect(setZos3270TerminalFolderExists).toHaveBeenCalledWith(false);
     });
   });
 });
